requirements.txt
python-dotenv
requests
langchain
langchain-core
langchain-community
tqdm
